Design MinStack
A MinStack is a special stack that supports standard stack operations (push, pop, top) and an additional operation getMin() which returns the minimum element in the stack in constant time O(1).

To achieve this, we use two stacks:

    The main stack stores all the elements.

    The min stack keeps track of the current minimum elements.

Every time we push or pop an element, we update both stacks accordingly. This way, the top of the min stack always holds the current minimum.